+++ pytorch_lightning/core/lightning.py
@@ -1610,11 +1610,12 @@ class LightningModule(
 
         return splits
 
-    def summarize(self, mode: Optional[str] = ModelSummary.MODE_DEFAULT) -> Optional[ModelSummary]:
+    def summarize(self, mode: Optional[str] = ModelSummary.MODE_DEFAULT,
+                  max_depth: Optional[int] = None) -> Optional[ModelSummary]:
         model_summary = None
 
         if mode in ModelSummary.MODES:
-            model_summary = ModelSummary(self, mode=mode)
+            model_summary = ModelSummary(self, mode=mode, max_depth=max_depth)
             log.info("\n" + str(model_summary))
         elif mode is not None:
             raise MisconfigurationException(f"`mode` can be None, {', '.join(ModelSummary.MODES)}, got {mode}")